Parse comments in any encoding
Comments are now properly parsed no matter the encoding
of the file or text.

This is ground work for multi-byte support in readr (UTF-16...)

+++ src/TokenizerFwf.cpp
@@ -26,7 +26,7 @@ skip_t skipComments(SourcePtr source) {
   return out;
 }
 
-std::vector<bool> emptyCols_(SourcePtr source, int n, SourceIterator begin) {
+std::vector<bool> emptyCols_(SourcePtr source, size_t n, SourceIterator begin) {
   SourceIterator end = source->end();
   std::vector<bool> is_white;
